Revert "refactor(core): Add fake navigation to primitives for code sharing (#59857)"
This reverts commit c2102c545bea25af0d18d062aa07a10303e6961b.
