+++ packages/common/testing/BUILD.bazel
@@ -10,7 +10,6 @@ ng_module(
     deps = [
         "//packages/common",
         "//packages/core",
-        "//packages/core/testing",
         "@npm//rxjs",
     ],
 )
