+++ packages/common/testing/src/private_export.ts
@@ -7,4 +7,3 @@
  */
 
 export {provideFakePlatformNavigation as ɵprovideFakePlatformNavigation} from './navigation/provide_fake_platform_navigation';
-export {FakeNavigation as ɵFakeNavigation} from './navigation/fake_navigation';
