--- packages/core/primitives/dom-navigation/BUILD.bazel
@@ -1,32 +1,0 @@
-load("//tools:defaults.bzl", "ts_library", "tsec_test")
-
-package(default_visibility = [
-    "//packages:__pkg__",
-    "//packages/common:__subpackages__",
-    "//packages/core:__subpackages__",
-    "//tools/public_api_guard:__pkg__",
-])
-
-ts_library(
-    name = "dom-navigation",
-    srcs = glob(
-        [
-            "*.ts",
-            "src/**/*.ts",
-        ],
-    ),
-)
-
-tsec_test(
-    name = "tsec_test",
-    target = "dom-navigation",
-    tsconfig = "//packages:tsec_config",
-)
-
-filegroup(
-    name = "files_for_docgen",
-    srcs = glob([
-        "*.ts",
-        "src/**/*.ts",
-    ]),
-)
