--- packages/core/primitives/dom-navigation/testing/BUILD.bazel
@@ -1,34 +1,0 @@
-load("//tools:defaults.bzl", "ts_library", "tsec_test")
-
-package(default_visibility = [
-    "//packages:__pkg__",
-    "//packages/common:__subpackages__",
-    "//packages/core/primitives/dom-navigation/testing:__subpackages__",
-    "//packages/core/testing:__subpackages__",
-    "//tools/public_api_guard:__pkg__",
-])
-
-ts_library(
-    name = "testing",
-    srcs = glob(
-        [
-            "**/*.ts",
-        ],
-    ),
-    deps = [
-        "//packages/core/primitives/dom-navigation",
-    ],
-)
-
-tsec_test(
-    name = "tsec_test",
-    target = "testing",
-    tsconfig = "//packages:tsec_config",
-)
-
-filegroup(
-    name = "files_for_docgen",
-    srcs = glob([
-        "*.ts",
-    ]),
-)
