--- packages/core/primitives/dom-navigation/testing/test/BUILD.bazel
@@ -1,30 +1,0 @@
-load("//tools:defaults.bzl", "jasmine_node_test", "karma_web_test_suite", "ts_library")
-
-ts_library(
-    name = "test_lib",
-    testonly = True,
-    srcs = glob(
-        ["**/*.ts"],
-    ),
-    # Visible to //:saucelabs_unit_tests_poc target
-    visibility = ["//:__pkg__"],
-    deps = [
-        "//packages/core/primitives/dom-navigation/testing",
-        "//packages/private/testing",
-    ],
-)
-
-jasmine_node_test(
-    name = "test",
-    bootstrap = ["//tools/testing:node"],
-    deps = [
-        ":test_lib",
-    ],
-)
-
-karma_web_test_suite(
-    name = "test_web",
-    deps = [
-        ":test_lib",
-    ],
-)
