+++ packages/core/testing/BUILD.bazel
@@ -13,7 +13,6 @@ ng_module(
         "//packages:types",
         "//packages/compiler",
         "//packages/core",
-        "//packages/core/primitives/dom-navigation/testing",
         "//packages/localize",
         "//packages/zone.js/lib:zone_d_ts",
         "@npm//@types/jasmine",
