--- packages/core/testing/src/testing_private_export.ts
@@ -1,9 +1,0 @@
-/**
- * @license
- * Copyright Google LLC All Rights Reserved.
- *
- * Use of this source code is governed by an MIT-style license that can be
- * found in the LICENSE file at https://angular.dev/license
- */
-
-export {FakeNavigation as ɵFakeNavigation} from '../../primitives/dom-navigation/testing';
